serverside/requirements.txt
whitenoise >= 6.0.0
gunicorn >= 20.1.0
djangorestframework >= 3.13.1